--- src/widgets/footer/ui/Footer.styles.ts
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-export const FooterContainer = styled.footer`
-  width: 100%;
-  height: 64px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
